pyproject.toml
[tool.poetry]
name = "ead-dados"
version = "0.1.0"
description = ""
authors = ["desenvolvimento3 <desenvolvimento3@novalar.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
